+++ src/components/common/Avatar/index.tsx
@@ -1,5 +1,4 @@
-import Image from 'next/image';
-import styled from '@emotion/styled';
+import * as Styled from './style';
 
 interface AvatarProps {
   iconName: string;
@@ -11,14 +10,8 @@ const Avatar = (props: AvatarProps) => {
   const { iconName, imageWidth = 68, imageHeight = 68, ...rest } = props;
 
   return (
-    <StyledAvatar src={`/assets/icons/${iconName}.svg`} alt="icon" width={imageWidth} height={imageHeight} {...rest} />
+    <Styled.Avatar src={`/assets/icons/${iconName}.svg`} alt="icon" width={imageWidth} height={imageHeight} {...rest} />
   );
 };
 
 export default Avatar;
-
-const StyledAvatar = styled(Image)`
-  align-self: center;
-  border-radius: 50%;
-  object-fit: cover;
-`;
