+++ src/components/common/Avatar/style.ts
@@ -1,0 +1,8 @@
+import Image from 'next/image';
+import styled from '@emotion/styled';
+
+export const Avatar = styled(Image)`
+  align-self: center;
+  border-radius: 50%;
+  object-fit: cover;
+`;
